+++ src/features/BookSearchPage/api/bestBookGetApi.ts
@@ -1,15 +1,6 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { aladinConfig } from '@shared/config/aladinConfig';
-
-// API 응답 타입 (이미지와 링크만 포함)
-export interface BestBookResponse {
-  item: Array<{
-    itemId: number;
-    cover: string; // 책 이미지 URL
-    link: string; // 책 상세 페이지 링크
-    title: string; // 책 제목
-  }>;
-}
+import { BestBookResponse } from '@features/BookSearchPage/types/types';
 
 export const bestBookGetApi = createApi({
   reducerPath: 'bestBookGetApi',
