+++ src/features/BookSearchPage/api/bookSearchApi.ts
@@ -1,38 +1,10 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { aladinConfig } from '@shared/config/aladinConfig';
-
-// 알라딘 상품 검색 Api 관련 인터페이스
-export interface BookResponse {
-  version: string;
-  totalResults: number;
-  item: Array<{
-    itemId: number; // 상품 고유 번호
-    title: string; // 제목
-    author: string; // 저자
-    cover: string; // 이미지
-    priceStandard: number; // 정가 가격
-    customerReviewRank: number; // 리뷰 점수
-    link: string; // 알라딘 페이지 링크
-  }>;
-}
-
-// 요청 파라미터 타입 정의
-export interface SearchBooksParams {
-  query: string; // 사용자 검색어
-  page: number; // 페이지 넘버
-  sort: string; // 정렬 기준
-}
-
-// 상품 조회 관련 인터페이스
-export interface RatingInfoResponse {
-  item: Array<{
-    subInfo?: {
-      ratingInfo?: {
-        ratingCount: number; // 평가 인원 수
-      };
-    };
-  }>;
-}
+import {
+  BookResponse,
+  SearchBooksParams,
+  RatingInfoResponse,
+} from '@features/BookSearchPage/types/types';
 
 export const bookSearchApi = createApi({
   reducerPath: 'bookSearchApi',
